MM: Enable the “Allows External Storage” option for binary data attribute
“Allows External Storage […] lets Core Data decide whether it stores binary data within SQLite or as external files, based on the size of the data. The underlying SQLite database can efficiently store binary data up to roughly 100 kilobytes directly in the database. This option should generally be enabled.

Keep in mind that including large binary data in your model objects will make them more expensive to keep in memory. If, in most cases, you need the binary data together with the other attributes in the entity, it makes sense to store them together. Otherwise, it might be worthwhile to store the binary data in a separate entity and create a relationship between the two.

Another alternative is to only store file names in Core Data and manage the storage of the actual data on disk yourself. However, you should have a very good reason to do this instead of just storing the data in Core Data itself. You’re taking on full responsibility to ensure data integrity between Core Data and your own binary storage, which isn’t always an easy task.”—Florian Kugler. “Core Data.”